% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{AD.cval}
\alias{AD.cval}
\title{Compute critical value for the Anderson-Darling test}
\usage{
AD.cval(n, prob)
}
\arguments{
\item{n}{Sample size.}

\item{prob}{Quantile.}
}
\description{
\code{KS.cval} returns the critical values of for the Anderson-Darling test statistic.
}

