% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{AD.stat}
\alias{AD.stat}
\title{Multivariate Anderson-Darling test statistic}
\usage{
AD.stat(x, nsum = 1000)
}
\arguments{
\item{x}{The sample.}

\item{nsum}{See \code{copula::polylog}.}
}
\description{
\code{AD.stat} returns the multivariate Anderson-Darling test statistic for uniformly distributed data.
}
\details{
\code{AD.stat} uses \code{polylog} from the \code{copula} package.
}
\examples{
set.seed(102015)
u <- matrix(runif(50), ncol=2)
plot(u)
AD.stat(u)
#[1] 1.570951

x <- matrix( seq(0.01, 0.99, length=100), ncol=2)
plot(x)
AD.stat(x)

x <- matrix(c(0.5,0.1), ncol=2)
plot(x)
AD.stat(x)

x <- MASS::mvrnorm(n = 200, rep(0, 2), diag(1,2,2))
plot(x)
u <- rosenblatt.norm(x, rep(0,2), diag(1,2,2))
plot(u)
AD.stat(u)

# true situation: Data from Gumbel copula with known corellation.
# false assumption: Gauss copula
# Rosenblatt#s transformation under wrong assumptions and computation of AD.stat
set.seed(5)
library("gumbel")
gdata <- qlnorm(rgumbel(50,alpha=2,dim=5))
pairs(gdata)
sigma2 <- matrix((1/sqrt(2)),nrow=5,ncol=5)-diag(((1/sqrt(2))-1),5)
sigma2
ndata <- qnorm(plnorm(gdata))
pairs(ndata)
rbdata <- rosenblatt.norm(ndata, rep(0,5),sigma2)
pairs(rbdata)
AD.stat(rbdata)
}
\seealso{
Other X.test.statistics.: \code{\link{CvM.stat}},
  \code{\link{KSa.stat}}, \code{\link{biKS.stat}}
}

