% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{CvM.cval}
\alias{CvM.cval}
\title{Compute critical value for the Cramér-von Mises test}
\usage{
CvM.cval(n, prob)
}
\arguments{
\item{n}{Sample size.}

\item{prob}{Quantile.}
}
\description{
\code{KS.cval} returns the critical values of for the Cramér-von Mises test statistic.
}

