% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{CvM.stat}
\alias{CvM.stat}
\title{Multivariate Cramér-von Mises statistic}
\usage{
CvM.stat(u)
}
\arguments{
\item{u}{Data set. Coloumn size equals the dimension of the samples and the line number equals the number of observations.}
}
\description{
\code{CvM.stat} returns the value of the Cramér-von Mises test statistic for testing uniformity.
}
\details{
The CvM test statistic is in fact the same as the squared L2 star discrepancy from \emph{Warnock (1972)}.
\code{u} has couloumn size equalto the dimension of the samples and line number equal to the number of observations.
}
\examples{
set.seed(102015)
u <- matrix(runif(50), ncol=2)
plot(u)
CvM.stat(u)
#[1] 0.1061887

x <- MASS::mvrnorm(n = 200, rep(0, 2), diag(1,2,2))
plot(x)
u <- rosenblatt.norm(x, rep(0,2), diag(1,2,2))
plot(u)
CvM.stat(u)

u <- 1/array(9:4, c(6,2))
CvM.stat(u)
}
\seealso{
\itemize{
 \item{Warnock, T.T., 1972. Computational investigations of low discrepancy point sets. In: Zaremba, S.K. (Ed.), Applications of Number Theory to Numerical Analysis. Academic Press, pp. 319-343.}
 \item{Chiu, S.N. and K.I. Liu. Generalized Cramér--von Mises goodness-of-fit tests for multivariate distributions. Computational Statistics & Data Analysis, 53(11):3817-3834, 2009.}
}

Other X.test.statistics.: \code{\link{AD.stat}},
  \code{\link{KSa.stat}}, \code{\link{biKS.stat}}
}

