% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{KS.cval}
\alias{KS.cval}
\title{Compute critical value for the Kolmogorov–Smirnov test}
\usage{
KS.cval(n, prob)
}
\arguments{
\item{n}{Sample size.}

\item{prob}{Quantile.}
}
\description{
\code{KS.cval} returns the critical values of for the Kolmogorov–Smirnov test statistic.
}

