% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{KS.p.value}
\alias{KS.p.value}
\title{# kritische Werte nach emp. Quantil bzw. Pearson
quantile(stats,probs=0.95)
qpearson(p=0.95,params=biz0ppar)}
\usage{
KS.p.value(n, alpha)
}
\description{
# kritische Werte nach emp. Quantil bzw. Pearson
quantile(stats,probs=0.95)
qpearson(p=0.95,params=biz0ppar)
}

