% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{KSa.stat}
\alias{KSa.stat}
\title{approximative Kolmogorov-Smirnov test statistic}
\usage{
KSa.stat(u)
}
\arguments{
\item{u}{Data set. Coloumn size equals the dimension of the samples and the line number equals the number of observations.}
}
\description{
\code{KSa.stat} returns the value of the approximative Kolmogorov-Smirnov test statistic
}
\details{
The statistic used by this function \eqn{D_n^*} from
Justel, A.; Peña, D.; Zamar, R. (1997). "A multivariate Kolmogorov–Smirnov test of goodness of fit". Statistics & Probability Letters 35 (3): 251–259. doi:10.1016/S0167-7152(97)00020-5.
as an approximation of the Kolmogorov-Smirnov test statistic.
}
\examples{
set.seed(102015)
u <- matrix(runif(50), ncol=2)
plot(u)
CvM.stat(u)
#[1] 0.006029274
KSa.stat(u)
#[1] 0.252292

x <- MASS::mvrnorm(n = 200, rep(0, 2), diag(1,2,2))
plot(x)
u <- rosenblatt.norm(x, rep(0,2), diag(1,2,2))
plot(u)
CvM.stat(u)
#[1] 0.002172406
biKS.stat(u)
#[1] 0.1058533

u <- 1/array(9:4, c(6,2))
CvM.stat(u)
KSa.stat(u)
}
\seealso{
\itemize{
 \item{Justel, A, D. Pe~na and R. Zamar, 1997. A multivariate Kolmogorov-Smirnov test of goodness of fit. Statistics & Probability Letters 35, 251-259.}
 \item{Chiu, S.N. and K.I. Liu. Generalized Cramér--von Mises goodness-of-fit tests for multivariate distributions. Computational Statistics & Data Analysis, 53(11):3817-3834, 2009.}
}

Other X.test.statistics.: \code{\link{AD.stat}},
  \code{\link{CvM.stat}}, \code{\link{biKS.stat}}
}

