% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{stat.MC}
\alias{stat.MC}
\title{Compute Monte-Carlo simulations of test statistic}
\usage{
stat.MC(n, rp, func)
}
\description{
\code{stat.MC} returns the simulated values of function \code{func}.
}
\details{
@examples
#KS.ntest.sims <- stat.MC(10,10,biKS.stat)

system.time(KS.n40.sims.1000plus <- stat.MC(40,9000,biKS.stat,start1000))
#save(".Random.seed",file="random_state_seed50.RData") ## save current state

quantile(KS.n12.sims,probs=(1-alphas))
devtools::use_data(KS.n100.sims, KS.n150.sims) 
save(".Random.seed",file="random_state_seed1001.RData") ## save current state
load("random_state_seed10e2.RData") ## restore state just after set.seed()
}

