% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{rosenblatt.norm}
\alias{rosenblatt.norm}
\title{Rosenblatt's transformation for normally distributed data}
\usage{
rosenblatt.norm(x, Mu, Sigma)
}
\arguments{
\item{x}{Array containing the data which have to be transformed. Each line represents a data vector with the coloumn number as dimension.}

\item{Mu}{Mean vector.}

\item{Sigma}{Correlation matrix.}
}
\description{
\code{rosenblatt.norm} returns Rosenblatt's transformation of normally distributed vectors which are the lines of the array \code{x}, with mean vector \code{Mu} and correlation matrix \code{Sigma}.
}
\examples{
x <- matrix(rnorm(50), nrow=25, ncol=2)
plot(x)
u <- rosenblatt.norm(x, rep(0,2), diag(1,2,2))
plot(u)

x <- MASS::mvrnorm(n = 500, rep(0, 2), diag(1,2,2))
plot(x)
u <- rosenblatt.norm(x, rep(0,2), diag(1,2,2))
plot(u)

set.seed(5)
library("gumbel")
gdata <- qlnorm(rgumbel(12,alpha=2,dim=5))
pairs(gdata)
sigma2 <- matrix((1/sqrt(2)),nrow=5,ncol=5)-diag(((1/sqrt(2))-1),5)
sigma2
ndata <- qnorm(plnorm(gdata))
pairs(ndata)
rbdata <- rosenblatt.norm(ndata, rep(0,5),sigma2)
pairs(rbdata)
}
\seealso{
\code{\link[copula]{rtrafo}} for Rosenblatt's transformation on elliptical and Archimedean copulas.
}

