% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGOV.R
\name{stat.MC}
\alias{stat.MC}
\title{Compute Monte-Carlo simulations of test statistic \code{func}}
\usage{
stat.MC(n, rp, func)
}
\arguments{
\item{n}{Sample size.}

\item{rp}{Number of Monte Carlo simulations.}

\item{func}{Function which evaluates the test statistic.}
}
\description{
\code{stat.MC} returns the simulated values of function \code{func}.
}
\examples{
KS.n12test.sims <- stat.MC(12,100,biKS.stat)
# now quantiles can be estimated
quantile(KS.n12test.sims,probs=(1-c(0.01,0.05,0.1,0.25)))
}

